+++ test/api/api_tests.h
@@ -11,6 +11,7 @@
 #include "call.h"
 #include "call_calls_foreign.h"
 #include "call_wren_call_root.h"
+#include "debug.h"
 #include "error.h"
 #include "get_variable.h"
 #include "foreign_class.h"