+++ test/api/debug.h
@@ -1,0 +1,3 @@
+#include "wren.h"
+
+WrenForeignMethodFn debugBindMethod(const char* signature);